Allow a default source to be configured via environment variable
Enable function-kcl to be used as a base image for custom functions built in KCL
by allowing the function input to be empty when a default source is set via
environment variable. A user can build a custom function on top of function-kcl
by including a directory containing a KCL module and setting the environment
variable `FUNCTION_KCL_DEFAULT_SOURCE` to the module's path.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>